+++ projects/net.splitcells.gel.core/CHANGELOG.md
@@ -21,6 +21,7 @@
   2. Rename `SupplySelector` to `SupplyOfflineSelector`.
   3. Rename `SupplySelectors` to `SupplyOfflineSelectors`.
 ### Minor Changes
+* **2022-07-24** **\#8**: Deprecate `LinearInitialization` and prefer using `OnlineLinearInitialization` for better performance.
 * **2022-06-03** **\#8**: The history interface for allocations defines now a `toAnalysisFods` method,
   that creates an alternative more searchable version of `Table#toFods`.
   The new version has dedicated columns for fields of complex attributes.